pyproject.toml
[project]
name = "django-app"
version = "0.1.0"
description = "Library Management API: Django APP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-environ>=0.13.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "markdown>=3.10.2",
    "psycopg2-binary>=2.9.11",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.15.2",
]

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = ["**/migrations/*.py"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = false
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
pretty = true
